+++ RELEASE_NOTES.md
@@ -1,5 +1,6 @@
 # Release Notes
 -   ctrl + s: save screenshot to clipboard
+-   trace view with mouse hover
 
 ###  0.2.0 - 2020-04-13
 -   option for transposed display (horizontal time-scale)